Unify extension points
