--- src/main/java/sonia/scm/webhook/DefaultWebHookExecutorProvider.java
@@ -1,55 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020-present Cloudogu GmbH and Contributors
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package sonia.scm.webhook;
-
-import com.cloudogu.scm.el.ElParser;
-import com.google.inject.Inject;
-import com.google.inject.Provider;
-import sonia.scm.plugin.Extension;
-import sonia.scm.repository.Changeset;
-import sonia.scm.repository.Repository;
-
-@Extension
-public class DefaultWebHookExecutorProvider implements WebHookExecutorProvider<SimpleWebHook> {
-
-  private final Provider<WebHookHttpClient> httpClientProvider;
-  private final ElParser elParser;
-
-  @Inject
-  public DefaultWebHookExecutorProvider(Provider<WebHookHttpClient> httpClientProvider, ElParser elParser) {
-    this.httpClientProvider = httpClientProvider;
-    this.elParser = elParser;
-  }
-
-  @Override
-  public boolean handles(Class<WebHook> webHookClass) {
-    return SimpleWebHook.class.isAssignableFrom(webHookClass);
-  }
-
-  @Override
-  public WebHookExecutor createExecutor(SimpleWebHook webHook, Repository repository, Iterable<Changeset> changesets) {
-    return new DefaultWebHookExecutor(httpClientProvider.get(), elParser, webHook, repository, changesets);
-  }
-}
