+++ src/main/java/sonia/scm/webhook/WebHookSpecification.java
@@ -25,8 +25,14 @@
 package sonia.scm.webhook;
 
 import sonia.scm.plugin.ExtensionPoint;
+import sonia.scm.repository.Changeset;
+import sonia.scm.repository.Repository;
 
 @ExtensionPoint
-public interface WebHookSpecification {
-  Class<? extends SingleWebHookConfiguration> getSpecificationType();
+public interface WebHookSpecification<T extends SingleWebHookConfiguration> {
+  Class<T> getSpecificationType();
+
+  boolean handles(Class<WebHook> webHookClass);
+
+  WebHookExecutor createExecutor(T webHook, Repository repository, Iterable< Changeset > changesets);
 }
